fixed: silence compiler warnings
git-svn-id: https://xbmc.svn.sourceforge.net/svnroot/xbmc/branches/AudioEngine@35699 568bbfeb-2a22-0410-94d2-cc84cf5bfa90

+++ xbmc/cores/AudioEngine/AEUtil.h
@@ -41,8 +41,8 @@ public:
   {
     static const double k = 0.9f;
     /* perform a soft clamp */
-         if (x >  k) x = tanh((x - k) / (1 - k)) * (1 - k) + k; 
-    else if (x < -k) x = tanh((x + k) / (1 - k)) * (1 - k) - k;
+         if (x >  k) x = (float) (tanh((x - k) / (1 - k)) * (1 - k) + k);
+    else if (x < -k) x = (float) (tanh((x + k) / (1 - k)) * (1 - k) - k);
 
     /* hard clamp anything still outside the bounds */
     if (x >  1.0f) return  1.0f;
